make turn on message more generic
