+++ frontend/components/GenericMsgWithNavButton/_styles.scss
@@ -1,0 +1,3 @@
+.generic-msg-with-nav-button {
+  max-width: 520px;
+}
