--- frontend/components/TurnOnMessage/_styles.scss
@@ -1,3 +1,0 @@
-.turn-on-message {
-  max-width: 520px;
-}
