+++ frontend/pages/ManageControlsPage/OSUpdates/OSUpdates.tsx
@@ -17,7 +17,7 @@ import PremiumFeatureMessage from "components/PremiumFeatureMessage";
 import SectionHeader from "components/SectionHeader";
 import Spinner from "components/Spinner";
 
-import TurnOnMessage from "../../../components/TurnOnMessage/TurnOnMessage";
+import GenericMsgWithNavButton from "../../../components/GenericMsgWithNavButton/GenericMsgWithNavButton";
 import CurrentVersionSection from "./components/CurrentVersionSection";
 import TargetSection from "./components/TargetSection";
 import { parseOSUpdatesCurrentVersionsQueryParams } from "./components/CurrentVersionSection/CurrentVersionSection";
@@ -117,7 +117,15 @@ const OSUpdates = ({ router, teamIdForApi, queryParams }: IOSUpdates) => {
     !config?.mdm.enabled_and_configured &&
     !config?.mdm.windows_enabled_and_configured
   ) {
-    return <TurnOnMessage router={router} />;
+    return (
+      <GenericMsgWithNavButton
+        header="Manage your hosts"
+        info="MDM must be turned on to change settings on your hosts."
+        path={PATHS.ADMIN_INTEGRATIONS_MDM}
+        buttonText="Turn on"
+        router={router}
+      />
+    );
   }
 
   // If the user has not selected a platform yet, we default to the platform that
