+++ frontend/pages/ManageControlsPage/SetupExperience/cards/SetupAssistant/SetupAssistant.tsx
@@ -10,11 +10,12 @@ import mdmAPI, {
   IAppleSetupEnrollmentProfileResponse,
 } from "services/entities/mdm";
 import { DEFAULT_USE_QUERY_OPTIONS } from "utilities/constants";
+import PATHS from "router/paths";
 
 import SectionHeader from "components/SectionHeader";
 import Spinner from "components/Spinner";
 import CustomLink from "components/CustomLink";
-import TurnOnMessage from "components/TurnOnMessage";
+import GenericMsgWithNavButton from "components/GenericMsgWithNavButton";
 
 import SetupAssistantProfileUploader from "./components/SetupAssistantProfileUploader";
 import SetupAssistantProfileCard from "./components/SetupAssistantProfileCard/SetupAssistantProfileCard";
@@ -128,10 +129,11 @@ const SetupAssistant = ({
       )
     ) {
       return (
-        <TurnOnMessage
+        <GenericMsgWithNavButton
           header="Additional configuration required"
           info="To customize, first turn on automatic enrollment."
           buttonText="Turn on"
+          path={PATHS.ADMIN_INTEGRATIONS_MDM}
           router={router}
         />
       );
